requirements.txt
# requirements.txt
alabaster==1.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bibtexparser==1.4.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
Deprecated==1.2.18
distro==1.9.0
docutils==0.21.2
exa-py==1.14.16
fake-useragent==2.2.0
free-proxy==1.1.3
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-search-results==2.4.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
MarkupSafe==3.0.2
narwhals==1.47.0
numpy==2.3.1
openai==1.96.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pdfminer.six==20250506
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.9.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
redis==5.3.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
roman-numerals-py==3.1.0
rpds-py==0.26.0
rsa==4.9.1
scholarly==1.7.11
selenium==4.34.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soupsieve==2.7
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
streamlit==1.46.1
tavily-python==0.7.9
tenacity==9.1.2
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
websocket-client==1.8.0
whitenoise==6.7.0
wrapt==1.17.2
wsproto==1.2.0
dj-database-url==2.2.0
gunicorn==22.0.0
psycopg[binary] # For PostgreSQL connectivity
django-allauth==0.63.0 # For user authentication and social login
django-redis==5.4.0 # For Redis cache backend (used for Render deployment)
